% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_09.R
\name{opt_mean1}
\alias{opt_mean1}
\title{Mean optimization}
\usage{
opt_mean1(x, fn, iter = 100)
}
\arguments{
\item{x}{A \code{matrix} data}

\item{fn}{A \code{string} containing the function to be optimized.}

\item{iter}{A \code{numeric} (integer) used to denote the number of iterations.}
}
\value{
A \code{number} containing the minimum of loss function:
}
\description{
Compute the minimum of mean.
}
\examples{
opt_mean1(rnrom(10, 2, 1), fn = loss)
}
\author{
Yagmur Yavuz Ozdemir
}
