% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_09.R
\name{opt_mean2}
\alias{opt_mean2}
\title{Mean optimization 2}
\usage{
opt_mean2(init = runif(1, min(x), max(x)), fn, x, iter = 100)
}
\arguments{
\item{init}{A \code{numeric} initial point}

\item{fn}{A \code{string} containing the function to be optimized.}

\item{x}{A \code{matrix} data}

\item{iter}{A \code{numeric} (integer) used to denote the number of iterations.}
}
\value{
A \code{number} containing the minimum of loss function:
}
\description{
Compute the minimum of mean.
}
\examples{
opt_mean2(init=runif(1, min(rnorm(10)), max(rnorm(10))) , fn = loss, rnorm(10) )
}
\author{
Yagmur Yavuz Ozdemir
}
